The second part of adding support for listing members of struct, class, and union.
Here we add support for listing base classes.
There is still some strangeness in listing classes, where fields with no name, type, or value appear.
Those are methods and the base class object.Object. I need to filter those out.

Now we handle looking up objects with a full dotted name, and then only do a child in parent search if the full name search fails.
Note that we still don't handle static data members. To fully handle them we need a change in DMD so that it stores the names of non-public symbols unmangled.