The third step in supporting listing member of struct, class, and union.
We now filter debug info for these types, so that only data members and base classes are shown.

Certain base classes will show up as a no-name child of a class. This includes object.Object. 
The problem is that they're not listed under the symbols section in the debug info, only in the types section.
Classes that you define in your program are stored in both places, and we look them up only under symbols.

Another thing missing is being able to detect what an object's real type is instead of only its declared type.
I haven't put the support for figuring that out. But, even so, the virtual table info used for doing that isn't being stored in the debug info right now, so it won't do any good.