Adding support for associative arrays (tables). For now we only allow keys of basic, enum, pointer, and delegate types.
Fixing a bug in the function that reads an array element ExprContext::GetValue. 
If the size of the element is bigger than our buffer, we only read part of the value. 
It's really an error case, so now we fail.

Fixing a bug in Utf8To16. It failed, because it called MultiByteToWideChar with a wrong flag.

Fixing a bug in comparing floating point values. 
The problem was that you could have a full float80 value in a DataValue, but the type could be float32 or float64. 
We're supposed to narrow the values to the common type of the comparison.